:star: HF_0000034: Product Uom Packagine New Column

+++ database/migrations/2025_04_10_162033_add_some_column_in_pos_terminal_transactions_table.php
@@ -35,7 +35,7 @@ class AddSomeColumnInPosTerminalTransactionsTable extends Migration
                     $table->tinyInteger('billing_type')->default(BillingType::COD)->after('order_schedule');
                 }
                 if (! Schema::hasColumn('pos_terminal_transactions', 'total_delivery_fee')) {
-                    $table->tinyInteger('total_delivery_fee')->default(0.000000)->after('total_discount_amount');
+                    $table->decimal('total_delivery_fee')->default(0.000000)->after('total_discount_amount');
                 }
             });
         }
